backend/requirements.txt
channels==4.3.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
paramiko==4.0.0
